fixed package names lowrCase
