+++ Config/Defines.java
@@ -1,4 +1,4 @@
-package Config;
+package config;
 
 /**
  * Defines
