--- Core/ORM/DataBaseCollection.java
@@ -1,8 +1,0 @@
-package Core.ORM;
-
-/**
- * DataBaseCollection
- */
-public class DataBaseCollection {
-
-}
