+++ examples/User.java
@@ -1,0 +1,23 @@
+package examples;
+
+import dopsie.core.*;
+import dopsie.exceptions.*;
+import java.util.ArrayList;
+
+/**
+ * User Model
+ */
+public class User extends Model {
+
+    public int id;
+
+    @Override
+    public String getTableName() {
+        return "personne";
+    }
+
+    @Override
+    public String getPrimaryKeyName() {
+        return "id";
+    }
+}
